News_Application/requirements.txt
alabaster==1.0.0
asgiref==3.9.2
babel==2.17.0
black==25.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
defusedxml==0.7.1
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-xml==2.0.0
djangorestframework_simplejwt==5.5.1
docutils==0.21.2
flake8==7.3.0
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
mariadb==1.1.14
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mypy_extensions==1.1.0
mysqlclient==2.2.7
myst-parser==5.0.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.4.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
pytokens==0.1.10
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
roman-numerals==4.1.0
roman-numerals-py==4.1.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
tweepy==4.16.0
tzdata==2025.2
urllib3==2.6.2
